backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ── Settings ───────────────────────────────────────────────────────────────────
pydantic==2.7.4
pydantic-settings==2.3.3

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30

# ── Task queue ─────────────────────────────────────────────────────────────────
celery==5.4.0
redis==5.0.7
flower==2.0.1

# ── File processing ────────────────────────────────────────────────────────────
Pillow==10.4.0
pdfplumber==0.11.0
pypdf==4.3.1

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
